+++ packages/mobile/src/components/Touchable/index.tsx
@@ -142,7 +142,7 @@ const _Touchable = forwardRef<
 
   }
 
-  const _styles = StyleSheet.flatten([variantStyles.wrapper, style])
+  const _styles = StyleSheet.flatten([variantStyles.wrapper, props?.disabled && variantStyles['wrapper:disabled'], style])
 
   const disableFeedback = !onPress || noFeedback