+++ packages/web/src/lib/utils/index.ts
@@ -1,4 +1,6 @@
 export * from './pollyfils/scroll'
 export * from './stopPropagation'
+export * from './useMaxContentWidth'
 export { default as url } from 'url-parse'
 export { default as Cookies } from './cookies'
+